Move shared recipe implementations to patch packages and fix YAML references
