+++ camel-upgrade-recipes/src/test/java/org/apache/camel/upgrade/camel418_3/RenameHeaderPrefixInXmlDslTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.upgrade.camel421;
+package org.apache.camel.upgrade.camel418_3;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
